# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "chekapps api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "User",
                "description": "User",
                "operationId": "user",
                "responses": {
                    "200": {
                        "description": "Successfully response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        },
                                        "user_fio": {
                                            "type": "string"
                                        },
                                        "user_email": {
                                            "type": "string"
                                        },
                                        "balance": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/check-histories": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Позволяет получить историю чеков пользователя",
                "description": "Позволяет получить список чеков пользователя. Параметры не обязательны, по умолчанию user_id - это текущий пользователь, а paginate - 20",
                "operationId": "check-histories",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Позволяет получить историю чеков определенного пользователя по его ID",
                        "required": false,
                        "example": "1"
                    },
                    {
                        "name": "paginate",
                        "in": "query",
                        "description": "Определяет количество чеков на одной странице",
                        "required": false,
                        "example": "3"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер запрашиваемой страницы",
                        "required": false,
                        "example": "1"
                    }
                ],
                "responses": {
                    "404": {
                        "description": "404 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Вылетает если запрос не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successfully response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number"
                                        },
                                        "user_id": {
                                            "type": "number"
                                        },
                                        "check_id": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "reward": {
                                            "type": "number"
                                        },
                                        "comment": {
                                            "type": "null|string"
                                        },
                                        "created_at": {
                                            "description": "Date this interaction was created",
                                            "type": "string",
                                            "format": "date-format"
                                        },
                                        "user": {
                                            "properties": {
                                                "check_id": {
                                                    "type": "number"
                                                },
                                                "image": {
                                                    "type": "string"
                                                },
                                                "amount": {
                                                    "type": "number"
                                                },
                                                "amount_in_currency": {
                                                    "type": "number"
                                                },
                                                "currency": {
                                                    "type": "string"
                                                },
                                                "verify_quantity": {
                                                    "type": "number"
                                                },
                                                "current_quantity": {
                                                    "type": "number"
                                                },
                                                "status": {
                                                    "type": "string"
                                                },
                                                "check_user_id": {
                                                    "type": "null|number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Позволяет получить всех пользователей с пагинацией",
                "description": "Позволяет получить всех пользователей с пагинацией. Параметры не обязательны, по умолчанию paginate = 20, а filter = DESC",
                "operationId": "users",
                "parameters": [
                    {
                        "name": "paginate",
                        "in": "query",
                        "description": "Определяет количество пользователей на одной странице",
                        "required": false,
                        "example": "20"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер запрашиваемой страницы",
                        "required": false,
                        "example": "1"
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Фильтр пользователей по активности DESC - самые активные, ASC - самые неактивные",
                        "required": false,
                        "example": "DESC"
                    },
                    {
                        "name": "searchBy",
                        "in": "query",
                        "description": "По какому полю искать пользователя user_fio - ФИО, user_email - email. Если будет использоваться поиск, то этот параметр обязателен",
                        "required": false,
                        "example": "user_fio"
                    },
                    {
                        "name": "s",
                        "in": "query",
                        "description": "Поисковый запрос",
                        "required": false,
                        "example": "Никита"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Вылетает если запрос не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Вылетает если у пользователя не прав для этого метода",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        },
                                        "code": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successfully response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number"
                                        },
                                        "user_id": {
                                            "type": "number"
                                        },
                                        "user_fio": {
                                            "type": "string"
                                        },
                                        "user_email": {
                                            "type": "string"
                                        },
                                        "user_phone": {
                                            "type": "string"
                                        },
                                        "balance": {
                                            "type": "number"
                                        },
                                        "check_history_count": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/users/{user_id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Позволяет получить определенного пользователя по user_id",
                "description": "Позволяет получить определенного пользователя по user_id",
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "example": "513753"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Вылетает если запрос не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Вылетает если у пользователя не прав для этого метода",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        },
                                        "code": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successfully response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number"
                                        },
                                        "user_id": {
                                            "type": "number"
                                        },
                                        "user_fio": {
                                            "type": "string"
                                        },
                                        "user_email": {
                                            "type": "string"
                                        },
                                        "user_phone": {
                                            "type": "string"
                                        },
                                        "balance": {
                                            "type": "number"
                                        },
                                        "isAdmin": {
                                            "type": "boolean"
                                        },
                                        "permissions": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "slug": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "checkHistory": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number"
                                                    },
                                                    "user_id": {
                                                        "type": "number"
                                                    },
                                                    "check_id": {
                                                        "type": "number"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "reward": {
                                                        "type": "number"
                                                    },
                                                    "comment": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "security": [
        []
    ]
}